Path Structure Support for Python AlternateCostsAlongPath
